+++ app/src/main/res/drawable/bg_home.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android"
+    android:shape="rectangle">
+
+    <solid android:color="@color/blue_backgound"/>
+    <corners android:topRightRadius="32dp"
+        android:topLeftRadius="24dp"/>
+
+</shape>
