--- app/src/main/res/drawable/blackboard.xml
@@ -1,72 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="90dp"
-    android:height="90dp"
-    android:viewportWidth="90"
-    android:viewportHeight="90">
-  <path
-      android:pathData="M4.2188,14.0625H85.7812V75.9375H4.2188V14.0625Z"
-      android:fillColor="#FFB431"/>
-  <path
-      android:pathData="M9.8438,19.6875H80.1562V70.3125H9.8438V19.6875Z"
-      android:fillColor="#426887"/>
-  <path
-      android:pathData="M11.8336,19.6875H9.8438V21.6773L4.2188,16.0523V14.0625H6.2086L11.8336,19.6875Z"
-      android:fillColor="#F99E0C"/>
-  <path
-      android:pathData="M85.7814,14.0625V16.0523L80.1564,21.6773V19.6875H78.1665L83.7915,14.0625H85.7814Z"
-      android:fillColor="#F99E0C"/>
-  <path
-      android:pathData="M9.8438,19.6875H80.1562V22.5H9.8438V19.6875Z"
-      android:fillColor="#3C6180"/>
-  <path
-      android:pathData="M9.8438,19.6875H12.6562V70.3125H9.8438V19.6875Z"
-      android:fillColor="#3C6180"/>
-  <path
-      android:pathData="M65.0339,69.6219L70.6574,59.7809L73.0988,61.176L67.4753,71.017L65.0339,69.6219Z"
-      android:fillColor="#E4D1B7"/>
-  <path
-      android:pathData="M50.8103,67.5H60.6541V70.3125H50.8103V67.5Z"
-      android:fillColor="#E4D1B7"/>
-  <path
-      android:pathData="M15.4688,25.3125H28.125V28.125H15.4688V25.3125Z"
-      android:fillColor="#D8C6AD"/>
-  <path
-      android:pathData="M15.4688,30.9375H18.2812V33.75H15.4688V30.9375Z"
-      android:fillColor="#D8C6AD"/>
-  <path
-      android:pathData="M21.0938,30.9375H28.125V33.75H21.0938V30.9375Z"
-      android:fillColor="#D8C6AD"/>
-  <path
-      android:pathData="M15.4688,36.5625H23.9062V39.375H15.4688V36.5625Z"
-      android:fillColor="#D8C6AD"/>
-  <path
-      android:pathData="M68.6889,55.4991L51.8139,28.7803C51.7044,28.607 51.5583,28.4596 51.3858,28.3487C51.2134,28.2377 51.0188,28.1658 50.8156,28.138C50.6124,28.1102 50.4057,28.1272 50.2097,28.1878C50.0138,28.2483 49.8335,28.351 49.6815,28.4885L27.5747,48.4898L27.5717,48.4924L20.1502,55.2073C19.9396,55.3979 19.7916,55.6479 19.7258,55.9242C19.6599,56.2006 19.6794,56.4904 19.7815,56.7555C19.8836,57.0206 20.0637,57.2486 20.2979,57.4094C20.5321,57.5702 20.8095,57.6562 21.0936,57.6563H67.4999C67.7515,57.6563 67.9985,57.5888 68.2151,57.4608C68.4317,57.3329 68.61,57.1492 68.7315,56.9288C68.8529,56.7085 68.913,56.4596 68.9055,56.2081C68.898,55.9567 68.8232,55.7118 68.6889,55.4991ZM54.8436,54.8438H52.0311V52.0313H54.8436V54.8438ZM24.7439,54.8438L28.4329,51.506C29.1463,52.4726 29.5312,53.6424 29.5311,54.8438H24.7439ZM30.5232,49.6148L49.2186,32.6999V54.8438H32.3436C32.343,52.9455 31.7016,51.1031 30.5232,49.6148ZM57.6561,54.8438V50.625C57.6561,50.2521 57.508,49.8944 57.2443,49.6306C56.9805,49.3669 56.6229,49.2188 56.2499,49.2188H52.0311V34.3913L64.9484,54.8438H57.6561Z"
-      android:fillColor="#D8C6AD"/>
-  <path
-      android:pathData="M30.9375,36.5625H33.75V39.375H30.9375V36.5625Z"
-      android:fillColor="#D8C6AD"/>
-  <path
-      android:pathData="M35.1562,60.4688H37.9688V63.2812H35.1562V60.4688Z"
-      android:fillColor="#D8C6AD"/>
-  <path
-      android:pathData="M63.2812,39.375H66.0938V42.1875H63.2812V39.375Z"
-      android:fillColor="#D8C6AD"/>
-  <path
-      android:pathData="M4.2188,70.3125H85.7812V75.9375H4.2188V70.3125Z"
-      android:fillColor="#FFB431"/>
-  <path
-      android:pathData="M4.2188,67.5H9.8438V70.3125H4.2188V67.5Z"
-      android:fillColor="#F9A614"/>
-  <path
-      android:pathData="M71.7188,63.2812H85.7812V70.3125H71.7188V63.2812Z"
-      android:fillColor="#DE5540"/>
-  <path
-      android:pathData="M80.1562,63.2812H85.7812V70.3125H80.1562V63.2812Z"
-      android:fillColor="#D6513C"/>
-  <path
-      android:pathData="M71.7188,67.5H85.7812V70.3125H71.7188V67.5Z"
-      android:fillColor="#E4D1B7"/>
-  <path
-      android:pathData="M80.1562,67.5H85.7812V70.3125H80.1562V67.5Z"
-      android:fillColor="#D8C6AD"/>
-</vector>
