--- app/src/main/res/drawable/book.xml
@@ -1,36 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="90dp"
-    android:height="90dp"
-    android:viewportWidth="90"
-    android:viewportHeight="90">
-  <path
-      android:pathData="M8.8759,81.0155L0.2639,32.1207C0.1807,31.6488 0.1289,31.1719 0.1087,30.6931C-0.0193,25.1497 3.9611,20.3613 9.4346,19.4741L26.0535,16.5414L25.6035,81.0155H8.8759Z"
-      android:fillColor="#35495E"/>
-  <path
-      android:pathData="M72.838,78.8742C72.5498,80.0706 71.5819,80.9837 70.3707,81.2018L21.4604,89.8294C18.3776,90.3675 15.2123,89.5521 12.7732,87.5913C10.3342,85.6306 8.8576,82.7144 8.7207,79.588C8.5985,74.288 12.2475,69.6439 17.4259,68.5087L31.1276,72.3569H31.1121L18.7759,74.5449C17.4001,74.7867 16.2051,75.6326 15.5198,76.8499C14.8345,78.0672 14.7311,79.5276 15.238,80.8294C16.1157,82.926 18.3448,84.1183 20.5759,83.6845L57.957,77.0897L72.838,78.8742Z"
-      android:fillColor="#2C3E50"/>
-  <path
-      android:pathData="M31.1121,72.3568L18.7759,74.5447C17.4001,74.7866 16.2051,75.6325 15.5198,76.8498C14.8345,78.0671 14.7311,79.5275 15.238,80.8292C16.1157,82.9258 18.3448,84.1182 20.5759,83.6844L57.9569,77.0896L31.1121,72.3568Z"
-      android:fillColor="#F9EAB0"/>
-  <path
-      android:pathData="M19.9026,51.4723L28.5255,2.5712C28.8342,0.8903 30.4358,-0.2309 32.1209,0.0543L80.6541,8.6121C86.1315,9.4999 90.115,14.2914 89.9877,19.8388C89.9666,20.3154 89.9148,20.7902 89.8326,21.2602L81.2096,70.1612L19.9026,51.4723Z"
-      android:fillColor="#E64C3C"/>
-  <path
-      android:pathData="M70.5073,16.2528C70.4162,16.2527 70.3254,16.245 70.2357,16.2295L35.0876,10.0319C34.2944,9.8363 33.7887,9.0592 33.9312,8.2548C34.0737,7.4503 34.8155,6.8941 35.6276,6.9828L70.7742,13.1726C71.5668,13.3115 72.1218,14.0333 72.0523,14.835C71.9829,15.6368 71.312,16.2523 70.5073,16.2528Z"
-      android:fillColor="#3F5C6C"/>
-  <path
-      android:pathData="M63.0466,22.8103C62.9555,22.8104 62.8646,22.8021 62.775,22.7855L39.8529,18.7494C39.0282,18.5822 38.4871,17.7881 38.6332,16.9594C38.7792,16.1306 39.5592,15.5694 40.3914,15.6941L63.315,19.7286C64.1123,19.8633 64.6725,20.5877 64.6023,21.3933C64.5321,22.1988 63.8551,22.8155 63.0466,22.8103Z"
-      android:fillColor="#3F5C6C"/>
-  <path
-      android:pathData="M66.2865,29.741L34.1891,24.071C33.345,23.9219 32.5398,24.4853 32.3906,25.3295L29.6906,40.614C29.5415,41.4581 30.105,42.2633 30.9491,42.4124L63.0465,48.0824C63.8906,48.2315 64.6958,47.6681 64.845,46.824L67.545,31.5395C67.6941,30.6953 67.1306,29.8901 66.2865,29.741Z"
-      android:fillColor="#ECF0F1"/>
-  <path
-      android:pathData="M89.9846,19.8311C89.9661,20.31 89.9143,20.787 89.8294,21.2587L81.2018,70.1535L69.238,66.507L76.1742,27.1863L77.8191,17.907L79.4949,8.4104L80.6587,8.6121C86.1297,9.5037 90.1078,14.2894 89.9846,19.8311Z"
-      android:fillColor="#C03A2B"/>
-  <path
-      android:pathData="M72.0311,57.507L23.4931,48.957C23.3137,48.9261 23.132,48.9105 22.95,48.9104C21.3417,48.9145 20.003,50.1466 19.8658,51.749C19.7285,53.3515 20.8382,54.7933 22.4224,55.0707L71.3173,63.6828C72.6919,63.9318 73.8832,64.7824 74.565,66.0017C75.2467,67.2211 75.3475,68.6814 74.8397,69.9828C73.9667,72.0748 71.741,73.2623 69.5173,72.8225L20.8086,64.2414C20.6293,64.2105 20.4476,64.195 20.2655,64.1949C18.6543,64.1909 17.3079,65.4207 17.1664,67.0258C17.0249,68.6308 18.1353,70.0772 19.7224,70.3552L68.6328,78.9673C71.7134,79.5074 74.8773,78.6947 77.3164,76.7369C79.7554,74.7792 81.2333,71.8659 81.3724,68.7414C81.4942,63.1917 77.5099,58.3999 72.0311,57.507Z"
-      android:fillColor="#802F34"/>
-  <path
-      android:pathData="M74.8396,69.9827C73.9667,72.0747 71.7409,73.2622 69.5172,72.8224L20.8086,64.2414L22.4224,55.0707L71.3172,63.6827C72.6919,63.9317 73.8832,64.7823 74.5649,66.0016C75.2467,67.221 75.3474,68.6813 74.8396,69.9827Z"
-      android:fillColor="#F9EAB0"/>
-</vector>
