--- app/src/main/res/drawable/keys.xml
@@ -1,30 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="51dp"
-    android:height="51dp"
-    android:viewportWidth="51"
-    android:viewportHeight="51">
-  <path
-      android:pathData="M35.5217,6.4138L29.5451,0.4373C29.2642,0.1574 28.8847,0 28.4883,0H1.4941C0.6694,0 0,0.6694 0,1.4941V22.6113C0,23.4361 0.6694,24.1055 1.4941,24.1055H13.5469L22.5117,20.3123L35.959,14.6237V7.4707C35.959,7.0743 35.8016,6.6947 35.5217,6.4138ZM25.5,9.1641H19.5234C18.6987,9.1641 18.0293,8.4947 18.0293,7.6699C18.0293,6.8452 18.6987,6.1758 19.5234,6.1758H25.5C26.3248,6.1758 26.9941,6.8452 26.9941,7.6699C26.9941,8.4947 26.3248,9.1641 25.5,9.1641Z"
-      android:fillColor="#FF637B"/>
-  <path
-      android:pathData="M35.959,7.4707V14.6237L22.5117,20.3123V9.1641H25.5C26.3248,9.1641 26.9941,8.4947 26.9941,7.6699C26.9941,6.8452 26.3248,6.1758 25.5,6.1758H22.5117V0H28.4883C28.8847,0 29.2642,0.1574 29.5451,0.4373L35.5217,6.4138C35.8016,6.6947 35.959,7.0743 35.959,7.4707Z"
-      android:fillColor="#E63950"/>
-  <path
-      android:pathData="M51,37.4024V41.6288C51,42.0282 50.8436,42.3908 50.5876,42.6577C50.3156,42.9446 49.9311,43.1229 49.5058,43.1229H40.9543C40.1286,43.1229 39.4453,42.4615 39.4453,41.6367V38.9473H36.713C35.8882,38.9473 35.2617,38.2002 35.2617,37.3745V33.7676L33.8741,32.4358C31.2305,33.3702 28.2671,33.2198 25.7549,31.9856C25.0148,31.622 24.709,30.7265 25.0726,29.9864L25.1025,17.347L25.1075,15.2124C24.7449,14.4703 25.0517,13.5758 25.7938,13.2132C29.8061,11.2509 34.6242,12.0557 37.7848,15.2154C40.6366,18.0672 41.554,22.3225 40.2152,26.0927L50.5577,36.3415C50.8406,36.6214 51,37.0039 51,37.4024Z"
-      android:fillColor="#E0F4FE"/>
-  <path
-      android:pathData="M51,37.4023V41.6288C51,42.0282 50.8436,42.3908 50.5876,42.6577L48.0117,40.1346L25.1025,17.347L25.1075,15.2124C24.7449,14.4703 25.0517,13.5758 25.7938,13.2132C29.8061,11.2509 34.6242,12.0557 37.7848,15.2154C40.6366,18.0672 41.554,22.3225 40.2153,26.0927L50.5577,36.3415C50.8406,36.6214 51,37.0039 51,37.4023Z"
-      android:fillColor="#BCDBFD"/>
-  <path
-      android:pathData="M22.5117,12.1523C16.7443,12.1523 12.0527,16.8439 12.0527,22.6113C12.0527,26.6465 14.4155,30.3052 18.0293,32.0234V33.9459L15.4783,36.4959C14.8956,37.0796 14.8956,38.0259 15.4783,38.6096L17.4107,40.541L15.4783,42.4724C14.8956,43.0562 14.8956,44.0024 15.4783,44.5862L21.4549,50.5627C21.7467,50.8546 22.1292,51 22.5117,51C22.8942,51 23.2767,50.8546 23.5686,50.5627L26.5569,47.5744C26.8368,47.2935 26.9941,46.914 26.9941,46.5176V32.0234C30.608,30.3052 32.9707,26.6465 32.9707,22.6113C32.9707,16.8439 28.2791,12.1523 22.5117,12.1523Z"
-      android:fillColor="#FFCC75"/>
-  <path
-      android:pathData="M32.9707,22.6113C32.9707,26.6465 30.608,30.3052 26.9941,32.0234V46.5176C26.9941,46.914 26.8368,47.2935 26.5569,47.5744L23.5686,50.5627C23.2767,50.8546 22.8942,51 22.5117,51V12.1523C28.2791,12.1523 32.9707,16.8439 32.9707,22.6113Z"
-      android:fillColor="#FF9D49"/>
-  <path
-      android:pathData="M24.0059,7.6699V19.623C24.0059,20.4478 23.3365,21.1172 22.5117,21.1172C21.687,21.1172 21.0176,20.4478 21.0176,19.623V7.6699C21.0176,6.8452 21.687,6.1758 22.5117,6.1758C23.3365,6.1758 24.0059,6.8452 24.0059,7.6699Z"
-      android:fillColor="#62739A"/>
-  <path
-      android:pathData="M24.0059,7.6699V19.623C24.0059,20.4478 23.3365,21.1172 22.5117,21.1172V6.1758C23.3365,6.1758 24.0059,6.8452 24.0059,7.6699Z"
-      android:fillColor="#48578A"/>
-</vector>
