--- app/src/main/res/drawable/logout.xml
@@ -1,31 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="51dp"
-    android:height="51dp"
-    android:viewportWidth="51"
-    android:viewportHeight="51">
-  <group>
-    <clip-path
-        android:pathData="M0,0h51v51h-51z"/>
-    <path
-        android:pathData="M34.4649,1.4949V43.53C34.4649,44.3666 33.8077,45.0242 32.9707,45.0242H21.0176C20.1806,45.0242 19.5234,44.3666 19.5234,43.53L1.5938,1.4949C1.5938,0.6579 2.2509,0.0007 3.0879,0.0007H32.9707C33.8077,0.0007 34.4649,0.6579 34.4649,1.4949Z"
-        android:fillColor="#07485E"/>
-    <path
-        android:pathData="M34.4648,1.4949V43.53C34.4648,44.3666 33.8077,45.0242 32.9707,45.0242H21.0176C20.1806,45.0242 19.5234,44.3666 19.5234,43.53L15.041,33.0711V0.0007H32.9707C33.8077,0.0007 34.4648,0.6579 34.4648,1.4949Z"
-        android:fillColor="#03232E"/>
-    <path
-        android:pathData="M21.6748,9.1151L3.7451,0.1502C3.2972,-0.0591 2.7291,-0.0591 2.3109,0.2101C1.8626,0.5086 1.5938,0.9868 1.5938,1.4949V43.5301C1.5938,44.1873 2.012,44.7554 2.6097,44.9344L15.041,49.0884L20.5394,50.9109C20.9795,51.0872 21.5433,50.9876 21.8841,50.7319C22.2728,50.4331 22.5117,49.9849 22.5117,49.5067V10.4598C22.5117,9.8921 22.1829,9.3839 21.6748,9.1151Z"
-        android:fillColor="#FF8859"/>
-    <path
-        android:pathData="M22.5117,10.4598V49.5066C22.5117,49.9849 22.2724,50.4331 21.8841,50.7319C21.5464,50.9852 20.9837,51.0891 20.5394,50.9109L15.041,49.0884V5.798L21.6748,9.115C22.1829,9.3839 22.5117,9.8921 22.5117,10.4598Z"
-        android:fillColor="#FF3939"/>
-    <path
-        android:pathData="M38.2426,32.8945C37.7566,32.635 37.4531,32.1287 37.4531,31.577V30.0829H26.9941C26.1681,30.0829 25.5,29.4144 25.5,28.5887V22.6121C25.5,21.7861 26.1681,21.118 26.9941,21.118H37.4531V19.6239C37.4531,19.0721 37.7566,18.5659 38.2426,18.306C38.727,18.048 39.3181,18.0772 39.7761,18.3807L48.7409,24.3573C49.1568,24.6343 49.4063,25.1012 49.4063,25.6004C49.4063,26.0993 49.1568,26.5662 48.7409,26.8432L39.7761,32.8198C39.3126,33.1291 38.7216,33.1517 38.2426,32.8945Z"
-        android:fillColor="#FF3939"/>
-    <path
-        android:pathData="M16.5352,22.6121V28.5886C16.5352,29.4252 15.878,30.0828 15.041,30.0828C14.2041,30.0828 13.5469,29.4252 13.5469,28.5886V22.6121C13.5469,21.7751 14.2041,21.1179 15.041,21.1179C15.878,21.1179 16.5352,21.7751 16.5352,22.6121Z"
-        android:fillColor="#FFF5F5"/>
-    <path
-        android:pathData="M16.5352,22.6121V28.5886C16.5352,29.4252 15.878,30.0828 15.041,30.0828V21.1179C15.878,21.1179 16.5352,21.7751 16.5352,22.6121Z"
-        android:fillColor="#EFE2DD"/>
-  </group>
-</vector>
