--- app/src/main/res/layout/activity_subject_management.xml
@@ -1,73 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".SelectSubjectActivity">
-
-    <EditText
-        android:id="@+id/pt_add_subject"
-        android:layout_width="344dp"
-        android:layout_height="50dp"
-        android:layout_marginTop="8dp"
-        android:ems="10"
-        android:hint="@string/add_new_subject"
-        android:inputType="text"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.492"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/textView" />
-
-    <ImageView
-        android:id="@+id/imageViewSubject"
-        android:layout_width="30dp"
-        android:layout_height="30dp"
-        android:layout_marginStart="24dp"
-        android:layout_marginTop="24dp"
-        android:background="?selectableItemBackgroundBorderless"
-        android:padding="4dp"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:srcCompat="@drawable/arrow"
-        app:tint="@color/black" />
-
-    <TextView
-        android:id="@+id/textView"
-        style="@style/titleTextStyle"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="24dp"
-        android:text="@string/subject_management"
-        app:layout_constraintBottom_toBottomOf="@+id/imageViewSubject"
-        app:layout_constraintStart_toEndOf="@+id/imageViewSubject"
-        app:layout_constraintTop_toTopOf="@+id/imageViewSubject" />
-
-    <androidx.recyclerview.widget.RecyclerView
-        android:id="@+id/rv_subject_list"
-        android:layout_width="0dp"
-        android:layout_height="0dp"
-        android:scrollbars="vertical"
-        android:layout_marginTop="8dp"
-        android:clipToPadding="false"
-        android:paddingBottom="16dp"
-        app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="1.0"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/btn_add_subject"
-        tools:itemCount="6"
-        tools:listitem="@layout/item_subject" />
-
-    <Button
-        android:id="@+id/btn_add_subject"
-        android:layout_width="137dp"
-        android:layout_height="50dp"
-        android:layout_marginTop="4dp"
-        android:text="@string/save"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/pt_add_subject" />
-
-</androidx.constraintlayout.widget.ConstraintLayout>
