+++ app/src/main/res/values/strings.xml
@@ -1,5 +1,5 @@
 <resources>
-    <string name="app_name">QuizForFun</string>
+    <string name="app_name">My Quiz</string>
     <string name="password">Password</string>
     <string name="login">Login</string>
     <string name="username">Username</string>
